WIP Refactors modules/Authentication.js and signIn/signUp calls
